Split async task factory tests
Move shared task-factory recording helpers and ingestion stubs into a
fixture module, and separate the ingestion fan-out metadata scenario into
its own focused test module.

Remove the file-level `too-many-lines` suppression from the task-factory
test module while keeping the split files below the configured limit.
